ioutil no longer requires config
len(passages) will now return the number of files rather than the number of actual passages that are iterated over. This is to allow using shuffle() to shuffle the list of files. Shuffling the actual passages is harder and would not be efficient since the same files would have to be opened more than once or kept open until all of its passages have been read.