Fix swap check for implicit nodes
Since the root node had a larger index than all terminals, the order was switched when creating an implicit child for it. Since the root starts as the leftmost node, its index should be the lowest.
Now the root has index 0 and all terminals have index identical to their position.