Do not use remote edges for reentrancy to terminals/punctuation
Instead of adding multiple occurrences of terminals as remote children
to the non-terminal parents, just add them as regular children.
The parser now accepts multiple parents to terminals/punctuation.
When creating the final passage, the recurring occurrences of the
terminals and punctuation nodes need special treatment to create them
only once.
This resolves the problem where in bel_pmid_1692_4534.16206 we had a
concept with a placeholder, which in training still hasn't had all its
terminal children attached (because some were remote), and the
placeholder was thus resolved incorrectly, and the valid argument check
was incorrect: it saw up-01 instead of increase-01.