Fix bug where model.actions.all and model.model.labels[0] were different
After the first training iteration, the model is backed up, finalized,
used to evaluate on the dev set, and then restored. However, NN models
just return themselves when finalized, not a copy of themselves.
So the finalized model copied the list of actions to the new list, but
the pre-finalized one didn't, even though its underlying model.labels
changed.
Now we're synchronizing the model labels with the actions when restoring
the model after evaluating it on the dev set, fixing the bug.