Change orphan node label in AMR to :orphan, and do not use one for UCCA
Since UCCA evaluation is by edge spans, it makes no sense to add an edge
for orphan nodes rather than just omitting the orphan node and then
connecting all its children to the root instead.
For AMR we do want to keep the node itself because evaluation is by
triples rather than spans. So we now check if the nodes has a label,
and if so, keep it with a special :orphan edge.