Add Label transition to label nodes, instead of labeling at creation
A node label can now be created later, when the parser thinks it has
enough information (including attached children etc.) to label the node
correctly. Until then, the label is None and constraints can't really
be checked.

This solves SOME of the problem introduced by squashing all the <t%d>
placeholders into a single <t> etc., where in cases like "Are
witnessing" where the witness-02 is mistakenly aligned to both words,
failed the parser due to invalid oracle transition, because only "Are"
was attached and so the predicate was be-02 for which some of the args
were invalid. This is fixed now because the oracle only chooses Label
when the node has all of its terminal children.

There is still a problem with another funny case: "makes sense" where
both are aligned to sense-02. During the parse it chooses <l>-02 and
replaces it with make-02, for which some of the arguments are invalid.