Fix bug when training on AMR but initializing first on another format
Got assertion errors or arithmetic exceptions in Eigen when calling
forward(), in the softmax node, because the output dimension was zero
for the node labels, because they were set to be identical to the "n"
feature, which was first initialized for another format in init_model()
and had a size of 0 because the other format didn't use node labels.

Now initializing the feature parameter list to only the features with
non-zero dim, so that for non-AMR formats the feature simply won't be
there, and then we initialize it just when we need it, with the correct
parameters.

This should also make feature extraction more efficient in general
because we don't need to extract features we won't use anyway.