Format-specific LSTM support
The default is still to have one shared LSTM and no format-specific one,
but to get a format-specific LSTM one can use the --hyperparams argument
and specify the number of LSTM layers for each format (i.e. axis).
The parametesr for the axis models (format-specific BiRNNs) are saved
to the same model data file by DyNet, but the parameter names are saved
in a different dictionary to the model JSON.
Backward compatibility is maintained, i.e. models saved previously are
still loadable.