Support loading vocabulary from file, to avoid loading spaCy
If passages are already pre-annotated, we don't need spaCy for features.
We might only need spaCy when loading the word vector files, because it
maps vectors to word forms and not by integer ID, whereas the saved
feature values are integer IDs. Specifying a pre-calculated vocabulary
file created with tupa.scripts.dump_vocab allows skipping spaCy for that
too.